Added 2 more charts

+++ Models/State.cs
@@ -8,7 +8,7 @@ namespace H1B_Analysis_Project.Models
     public class State
     {
         public string StateName { get; set; }
-        public int NumberOfCertifiedCases { get; set; }
-        public float AverageWage { get; set; }
+        public double NumberOfCertifiedCases { get; set; }
+        public double AverageWage { get; set; }
     }
 }
